remove redundant unsafe
