+++ service/effector/src/main.rs
@@ -26,4 +26,3 @@ pub fn main() {
 
 #[cfg(not(target_arch = "wasm32"))]
 pub fn main() {}
-
